+++ config/routes.rb
@@ -11,16 +11,6 @@ DefraRubyMocks::Engine.routes.draw do
       as: "company_officers",
       constraints: ->(_request) { DefraRubyMocks.configuration.enabled? }
 
-  post "/worldpay/payments-service",
-       to: "worldpay#payments_service",
-       as: "worldpay_payments_service",
-       constraints: ->(_request) { DefraRubyMocks.configuration.enabled? }
-
-  get "/worldpay/dispatcher",
-      to: "worldpay#dispatcher",
-      as: "worldpay_dispatcher",
-      constraints: ->(_request) { DefraRubyMocks.configuration.enabled? }
-
   post "/govpay/v1/payments",
        to: "govpay#create_payment",
        as: "govpay_create_payment",
@@ -29,5 +19,15 @@ DefraRubyMocks::Engine.routes.draw do
   get "/govpay/v1/payments/:payment_id",
       to: "govpay#payment_details",
       as: "govpay_payment_details",
+      constraints: ->(_request) { DefraRubyMocks.configuration.enabled? }
+
+  post "/govpay/v1/payments/:payment_id/refunds",
+       to: "govpay#create_refund",
+       as: "govpay_create_refund",
+       constraints: ->(_request) { DefraRubyMocks.configuration.enabled? }
+
+  get "/govpay/v1/payments/:payment_id/refunds/:refund_id",
+      to: "govpay#refund_details",
+      as: "govpay_refund_details",
       constraints: ->(_request) { DefraRubyMocks.configuration.enabled? }
 end