+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## [Unreleased](https://github.com/DEFRA/defra-ruby-mocks/tree/HEAD)
+
+[Full Changelog](https://github.com/DEFRA/defra-ruby-mocks/compare/v2.4.1...HEAD)
+
+**Fixed bugs:**
+
+- Fix parameter handling in GovpayPaymentCallbackJob [\#118](https://github.com/DEFRA/defra-ruby-mocks/pull/118) ([PaulDoyle-DEFRA](https://github.com/PaulDoyle-DEFRA))
+
+## [v2.4.1](https://github.com/DEFRA/defra-ruby-mocks/tree/v2.4.1) (2023-05-23)
+
+[Full Changelog](https://github.com/DEFRA/defra-ruby-mocks/compare/v2.4.0...v2.4.1)
+
+**Fixed bugs:**
+
+- Fix govpay refund id handling [\#116](https://github.com/DEFRA/defra-ruby-mocks/pull/116) ([PaulDoyle-DEFRA](https://github.com/PaulDoyle-DEFRA))
+
+**Merged pull requests:**
+
+- Release/v2.4.1 [\#117](https://github.com/DEFRA/defra-ruby-mocks/pull/117) ([PaulDoyle-DEFRA](https://github.com/PaulDoyle-DEFRA))
+
+## [v2.4.0](https://github.com/DEFRA/defra-ruby-mocks/tree/v2.4.0) (2023-05-22)
+
+[Full Changelog](https://github.com/DEFRA/defra-ruby-mocks/compare/v2.3.3...v2.4.0)
+
+**Implemented enhancements:**
+
+- Add support for govpay refunds [\#112](https://github.com/DEFRA/defra-ruby-mocks/pull/112) ([PaulDoyle-DEFRA](https://github.com/PaulDoyle-DEFRA))
+
+**Merged pull requests:**
+
+- Version v2.4.0 [\#115](https://github.com/DEFRA/defra-ruby-mocks/pull/115) ([PaulDoyle-DEFRA](https://github.com/PaulDoyle-DEFRA))
+- Remove redundant Worldpay mocks [\#110](https://github.com/DEFRA/defra-ruby-mocks/pull/110) ([PaulDoyle-DEFRA](https://github.com/PaulDoyle-DEFRA))
+- Release 2.3.3 [\#108](https://github.com/DEFRA/defra-ruby-mocks/pull/108) ([PaulDoyle-DEFRA](https://github.com/PaulDoyle-DEFRA))
+
 ## [v2.3.3](https://github.com/DEFRA/defra-ruby-mocks/tree/v2.3.3) (2022-06-29)
 
 [Full Changelog](https://github.com/DEFRA/defra-ruby-mocks/compare/v2.3.2...v2.3.3)